Update build script.

+++ docker/registry/build.sh
@@ -19,19 +19,26 @@ docker cp builder:/go/bin/registry $TEMPDIR/registry/registry
 docker cp builder:/go/src/github.com/docker/distribution/cmd/registry/config-example.yml $TEMPDIR/registry/config-example.yml
 
 # Build the image.
-if [ ! -z ${IMAGE_TAG+x} ] ; then
-
-  echo "Building $IMAGE_NAME:$IMAGE_TAG..."
-  docker build --no-cache -t $IMAGE_NAME:$IMAGE_TAG .
-
-fi
-
 echo "Building $IMAGE_NAME:latest..."
+
 docker build --no-cache -t $IMAGE_NAME:latest .
 
-# Remove the builder.
-echo "Cleaning up..."
+# Make sure the last command completed successfully.
+if [ "$?" -eq 0 ]; then
+  
+  if [ ! -z ${IMAGE_TAG+x} ] ; then
+    echo "Building $IMAGE_NAME:$IMAGE_TAG..."
+    docker build --no-cache -t $IMAGE_NAME:$IMAGE_TAG .
+  fi
 
-rm -rf $TEMPDIR
-docker rm -f builder
-docker rmi $IMAGE_NAME:build
+  # Remove the builder.
+  echo "Cleaning up..."
+
+  rm -rf $TEMPDIR
+  docker rm -f builder
+  docker rmi $IMAGE_NAME:build
+
+else
+  exit "$?"
+
+fi
